requirements.txt
flask
jsonify
tensorflow
tensorflow_hub
pillow